stbt.conf: Change default confirm_method to normed-absdiff
This has been our recommended setting for over 4 years so it's a bit
unfortunate that new users should pay with an inferior algorithm for the
sake of backwards compatibility.

This is already the default value on the official [stb-tester test-pack]
which we use as a starting point for our customers that use the
stb-tester hardware.

[stb-tester test-pack]: https://github.com/stb-tester/stb-tester-test-pack