stbt.wait_until: Return the first stable value, not the last
If `stable_secs` is specified, wait_until should return the first stable
value. This is similar to how `wait_for_motion` reports the timestamp of
the first frame where motion was detected, even when you've asked to
validate continuous motion for several frames.

This allows uses like waiting for a selection to finish moving, and then
report how long the animation took.